1) Increased RAM and time efficiency. 2) Fixed DFS code. 3) Added comments. 4) Added new maze plot function. 5) Added functionality to ease answering questions.

+++ maze_generator.py
@@ -7,13 +7,12 @@ import random
 import numpy as np
 
 def maze_generator(maze_dimension, maze_probability):
-    data = np.zeros((maze_dimension, maze_dimension), dtype = int)
+    data = np.zeros((maze_dimension, maze_dimension), dtype = np.int8)
 
     for i in range(maze_dimension):
         for j in range(maze_dimension):
             if random.random() <= maze_probability:
                 data[i][j] = 1
-            
     data[0][0] = 0
     data[maze_dimension-1][maze_dimension-1] = 0
     return data
